Fix bug: parse error creates no output file
This commit fixes a problem where an input file which has a parse error
(for example, a 'compile' card of $PYTHON, not $PY) gives an error on
the console, but does not create an output log file so the user does not
see the error and can't correct it.

In studying this code I realized that it still has a problem (which I
did *NOT* address yet) wherein if the deck ends without supplying all
the expected control cards, no error is indicated on the log nor does it
create an output file, so again the user has no idea what the problem
is. Added a 'FIXME' comment for that but did not address it at this
time.