Further enhanced the parser to always create an output file
This commit reworks some of the parse logic to further reduce the case
where various types of input file construction errors would result in
the program not creating an output file (thus rendering the user unable
to even see the error).

In order to make this work I had to do away with the feature where it
was previously possible to have multiple jobs in one file (it was
getting to complicated to have it do both).